+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `.color-renvert` utility rule for reverting the color scheme
 - Neutral color palette
 - Dark and light color schemes
+- `.hue-<{ a, b, c, d }>` utility rules for applying the context of a hue palette
+- `.chroma-fg-<{ 1, 2, 3 }>` utility rules for applying the chroma of an active hue context to a foreground
+- `.chroma-bg-<{ 1, 2, 3 }>` utility rules for applying the chroma of an active hue context to a background
 
 ### Changed
 
@@ -28,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - `.tab-bar` component rule
+- `.solid` utility rule
 
 ### Fixed